requirements.txt
Flask==3.0.3
langchain==0.2.12
langchain_community==0.2.11
playsound==1.3.0
python-dotenv==1.0.1
Requests==2.32.3
